# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Music API",
    "description": "A comprehensive REST API for managing artists, albums, songs, and playlists with full CRUD operations, authentication, and data validation",
    "version": "2.0.0",
    "contact": {
      "name": "Music API Support",
      "email": "support@musicapi.com"
    }
  },
  "servers": [
    {
      "url": "https://music-api-cse341-u669.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/artists": {
      "get": {
        "tags": ["Artists"],
        "summary": "Get all artists with filtering, sorting, and pagination",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "description": "Filter by genre",
            "schema": { "type": "string" }
          },
          {
            "name": "country",
            "in": "query", 
            "description": "Filter by country",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page (max 100)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100 }
          }
        ],
        "responses": {
          "200": { "description": "Successfully retrieved artists" },
          "400": { "description": "Invalid query parameters" },
          "500": { "description": "Internal server error" }
        }
      },
      "post": {
        "tags": ["Artists"],
        "summary": "Create a new artist",
        "security": [{ "OAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ArtistInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Artist created successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Authentication required" },
          "409": { "description": "Artist already exists" }
        }
      }
    },
    "/artists/{id}": {
      "get": {
        "tags": ["Artists"],
        "summary": "Get artist by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Artist found" },
          "400": { "description": "Invalid ID format" },
          "404": { "description": "Artist not found" }
        }
      },
      "put": {
        "tags": ["Artists"],
        "summary": "Update artist",
        "security": [{ "OAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ArtistInput" }
            }
          }
        },
        "responses": {
          "204": { "description": "Artist updated successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Authentication required" },
          "404": { "description": "Artist not found" }
        }
      },
      "delete": {
        "tags": ["Artists"],
        "summary": "Delete artist",
        "security": [{ "OAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Artist deleted successfully" },
          "400": { "description": "Cannot delete artist with albums" },
          "401": { "description": "Authentication required" },
          "404": { "description": "Artist not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtistInput": {
        "type": "object",
        "required": ["name", "genre", "country", "formed_year", "members"],
        "properties": {
          "name": { "type": "string", "example": "The Beatles" },
          "genre": { "type": "string", "example": "Rock" },
          "country": { "type": "string", "example": "United Kingdom" },
          "formed_year": { "type": "integer", "example": 1960 },
          "members": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["John Lennon", "Paul McCartney", "George Harrison", "Ringo Starr"]
          },
          "biography": { "type": "string", "example": "English rock band formed in Liverpool." },
          "website": { "type": "string", "example": "https://www.thebeatles.com" },
          "social_media": {
            "type": "object",
            "example": { "twitter": "@thebeatles", "instagram": "@thebeatles" }
          }
        }
      },
...
    },
    "securitySchemes": {
      "OAuth": {
        "type": "oauth2",
        "description": "OAuth 2.0 authentication with Google",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/auth/google",
            "tokenUrl": "/auth/google/callback",
            "scopes": {
              "profile": "Access user profile",
              "email": "Access user email"
            }
          }
        }
      }
    }
  }
}